separte log_image tests
